Added new items:
Glob of Infection
Block of Infection
Infected Steak
Cassette Tape One

Current Issues:

Cassette Tape One is not playing in the jukebox.
Using the playsound command works just well and the audio is there. I can hear it, and there is no issues other than it not playing in the jukebox.